+++ spec/cursor.spec.js
@@ -94,5 +94,35 @@ describe('Cursor', function() {
       });
     });
 
+    describe('before()', function() {
+
+      it('gets the content before', function() {
+        var fragment = this.cursor.before();
+        expect(content.getInnerHtmlOfFragment(fragment)).toEqual('foobar');
+      });
+    });
+
+    describe('beforeHtml()', function() {
+
+      it('gets the content before', function() {
+        expect(this.cursor.beforeHtml()).toEqual('foobar');
+      });
+    });
+
+    describe('after()', function() {
+
+      it('gets the content after', function() {
+        var fragment = this.cursor.after();
+        expect(content.getInnerHtmlOfFragment(fragment)).toEqual('');
+      });
+    });
+
+    describe('afterHtml()', function() {
+
+      it('gets the content before', function() {
+        expect(this.cursor.afterHtml()).toEqual('');
+      });
+    });
+
   });
 });